Windows: stop supplying BLK_SHA1=YesPlease by default
Using BLK_SHA1 in lieu of the OpenSSL routines was done in
9bccfcdbff ("Windows: use BLK_SHA1 again", 2009-10-22). Since
DC_SHA1 is now the default for git in general, it makes sense for
Windows to use that too. This looks like something that was missed
back in e6b07da278 ("Makefile: make DC_SHA1 the default",
2017-03-17).

As noted in 2cfc70f0de ("mingw: use OpenSSL's SHA-1 routines",
2017-02-09) OpenSSL has a performance benefit compared to BLK_SHA1 on
MinGW, so perhaps that and the Windows default should be changed
around again. That's a topic for another series, it seems clear that
this specific flag is nobody's explicit intention.

Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>